refactor: improvements on ArcButton
